Add files via upload
Dictionary App
This is a simple dictionary web application that allows users to search for the meanings, examples, and synonyms of words using the Dictionary API.

Features
Search for Words: Users can enter a word in the search bar and click the search button to fetch the word's details.
Word Details: Displays the word, its meaning, example usage, and synonyms.
Responsive Design: The app is responsive and works well on different screen sizes.
Technologies Used
HTML: For the structure of the web page.
CSS: For styling the web page.
JavaScript: For fetching data from the API and updating the DOM.


How to Use
Clone the repository:
git clone https://github.com/Shashank80000/dictionary-app.git
Open index.html in your web browser.
Enter a word in the search bar and click the "Search" button to see the word's details.
Files
index.html: The main HTML file that contains the structure of the web page.
style.css: The CSS file that contains the styles for the web page.
script.js: The JavaScript file that contains the logic for fetching data from the API and updating the DOM.
License
This project is licensed under the MIT License. See the LICENSE file for more details.